Remove Thoughtbot dotfiles
Instead of blending my dotfiles with the Thoughtbot dotfiles, I've
decided to take them over myself. The downside of this is I no longer
get the awesome improvements that happen occasionally. However,
overwriting settings I didn't like was also tedious, and it was a pain
searching for settings in multiple places.