Fix the LSP config
This was a huge pain. I'm still not sure why my diagnostics quit
working, but everything seems to work well with these settings.

The downside to this configuration approach is that I have to manually
specify the language servers and null-ls sources I'd like to use, but
it's worth it to have everything work seemlessly.